requirements.txt
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
tqdm>=4.65.0
supabase
requests>=2.31.0
rapidfuzz>=3.0.0
